+++ internal/core/edit.go
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"os"
 	"os/exec"
+	"strings"
 
 	"github.com/sirupsen/logrus"
 
@@ -34,10 +35,8 @@ func Edit(uri string, dropTags bool, tags []secret.Tag, ignoreSchema bool) {
 				"Unable to decrypt secret! Are you allowed to access it?")
 			os.Exit(1)
 		}
-		output.Close()
 	case *secret.DoesNotExist:
 		s = secret.New(uri)
-		s.SetTags(tags)
 		if template := s.GetTemplate(); template != nil {
 			if err := ioutil.WriteFile(output.Name(), template.Skeleton, 0644); err != nil {
 				logrus.WithFields(logrus.Fields{
@@ -52,47 +51,44 @@ func Edit(uri string, dropTags bool, tags []secret.Tag, ignoreSchema bool) {
 			"uri":   uri,
 		}).Fatal("Failed to load secret!")
 	}
+	output.Close()
 
-	// Decide new tags.
-	var updateTags bool
-	var newTags []secret.Tag
+	// Compute initial digests.
+	tagsDigest := md5Tags(s)
+	secretDigest := md5File(output.Name())
+
+	// Adjust tags?
 	if dropTags {
-		updateTags = true
-		newTags = tags
-	} else {
-		updateTags = false
-		newTags = s.GetTags()
+		s.SetTags(make([]secret.Tag, 0))
+	} else if len(tags) > 0 {
+		s.SetTags(tags)
 	}
 
-	// Compute initial digest.
-	digest := md5File(output.Name())
-
 	// Avoid loosing edited changes.
+loop:
 	for {
-		// Open decrypted secret / initial skeleton in an external editor.
-		cmd := exec.Command(config.GetEditor(), output.Name())
-		cmd.Stdin = os.Stdin
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			logrus.WithFields(logrus.Fields{
-				"editor": config.GetEditor(),
-				"error":  err,
-			}).Fatal("Failed to open external editor!")
-		}
-
-		// Check if secret needs to be updated.
-		if updateTags || digest != md5File(output.Name()) {
-			if set(uri, output.Name(), newTags, ignoreSchema) {
+		openEditor(output.Name())
+		if tagsDigest != md5Tags(s) || secretDigest != md5File(output.Name()) {
+			if set(s, output.Name(), ignoreSchema) {
 				fmt.Println("Done!")
 				break
 			} else {
-				fmt.Print("\nReopen editor? (Y/n) ")
-				var response string
-				_, err := fmt.Scanln(&response)
-				if err == nil && response == "n" {
-					fmt.Println("Aborted!")
-					break
+				for {
+					fmt.Print("\nWhat now? edit [t]ags; edit [s]ecret; [a]bort ")
+					var response string
+					_, err := fmt.Scanln(&response)
+					if err == nil {
+						switch response {
+						case "t":
+							editTags(s)
+							continue loop
+						case "s":
+							continue loop
+						case "a":
+							fmt.Println("Aborted!")
+							break loop
+						}
+					}
 				}
 			}
 		} else {
@@ -100,6 +96,70 @@ func Edit(uri string, dropTags bool, tags []secret.Tag, ignoreSchema bool) {
 			break
 		}
 	}
+}
+
+func editTags(s *secret.Secret) {
+	// Initialize output writer.
+	output, err := ioutil.TempFile("", "pgp-tomb-")
+	if err != nil {
+		logrus.WithFields(logrus.Fields{
+			"error": err,
+		}).Fatal("Failed to open temporary file!")
+	}
+	defer os.Remove(output.Name())
+
+	// Dump & edit tags.
+	for _, tag := range s.GetTags() {
+		if _, err := fmt.Fprintf(output, "%s: %s\n", tag.Name, tag.Value); err != nil {
+			logrus.WithFields(logrus.Fields{
+				"file": output.Name(),
+				"error":  err,
+			}).Fatal("Failed to dump tags!")
+		}
+	}
+	output.Close()
+	openEditor(output.Name())
+
+	// Process new tags.
+	items, err := ioutil.ReadFile(output.Name())
+	if err != nil {
+		logrus.WithFields(logrus.Fields{
+			"file": output.Name(),
+			"error":  err,
+		}).Fatal("Failed to load tags!")
+	}
+	tags := make([]secret.Tag, 0)
+	for _, item := range strings.Split(string(items), `\n`) {
+		if index := strings.Index(item, ":"); index > 0 {
+			tags = append(tags, secret.Tag{
+				Name:  strings.TrimSpace(item[:index]),
+				Value: strings.TrimSpace(item[index+1:]),
+			})
+		}
+	}
+	s.SetTags(tags)
+}
+
+func openEditor(path string) {
+	cmd := exec.Command(config.GetEditor(), path)
+	cmd.Stdin = os.Stdin
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	if err := cmd.Run(); err != nil {
+		logrus.WithFields(logrus.Fields{
+			"editor": config.GetEditor(),
+			"error":  err,
+		}).Fatal("Failed to open external editor!")
+	}
+}
+
+func md5Tags(s *secret.Secret) string {
+	digest := md5.New()
+	for _, tag := range s.GetTags() {
+		digest.Write([]byte(tag.Name))
+		digest.Write([]byte(tag.Value))
+	}
+	return fmt.Sprintf("%x", digest.Sum(nil))
 }
 
 func md5File(path string) string {