+++ internal/core/set.go
@@ -8,21 +8,22 @@ import (
 
 	"github.com/sirupsen/logrus"
 
+	"github.com/carlosabalde/pgp-tomb/internal/core/config"
 	"github.com/carlosabalde/pgp-tomb/internal/core/secret"
 )
 
 func Set(uri, inputPath string, tags []secret.Tag, ignoreSchema bool) {
-	if !set(uri, inputPath, tags, ignoreSchema) {
-		os.Exit(1)
-	}
-	fmt.Println("Done!")
-}
-
-func set(uri, inputPath string, tags []secret.Tag, ignoreSchema bool) bool {
-	// Initializations.
 	s := secret.New(uri)
 	s.SetTags(tags)
 
+	if !set(s, inputPath, ignoreSchema) {
+		os.Exit(1)
+	}
+
+	fmt.Println("Done!")
+}
+
+func set(s *secret.Secret, inputPath string, ignoreSchema bool) bool {
 	// Initialize input reader.
 	var input io.Reader
 	if inputPath == "" {
@@ -38,6 +39,24 @@ func set(uri, inputPath string, tags []secret.Tag, ignoreSchema bool) bool {
 		}
 		defer file.Close()
 		input = file
+	}
+
+	// Check tags?
+	if !ignoreSchema {
+		if serializedTags, err := s.GetSerializedTags(); err != nil {
+			logrus.WithFields(logrus.Fields{
+				"error": err,
+			}).Error("Failed to serialize tags!")
+			return false
+		} else {
+			if valid, errs := validateSchema(serializedTags, config.GetTags()); !valid {
+				fmt.Fprintf(os.Stderr, "Tags do not match schema!\n")
+				for _, err := range errs {
+					fmt.Fprintf(os.Stderr, "  - %s\n", err)
+				}
+				return false
+			}
+		}
 	}
 
 	// Check template?